Старался навести порядок в коде
