+++ ImageFeed/Profile/ProfileStructures.swift
@@ -38,15 +38,3 @@ extension Profile {
         )
     }
 }
-
-struct UserResult: Codable {
-    let profileImage: ProfileImage
-    
-    private enum CodingKeys: String, CodingKey {
-        case profileImage = "profile_image"
-    }
-}
-
-struct ProfileImage: Codable {
-    let large: String
-}
