+++ ImageFeed/SceneDelegate.swift
@@ -21,10 +21,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let scene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: scene)
-//        window?.rootViewController = UIStoryboard(
-//            name: "Main",
-//            bundle: .main
-//        ).instantiateInitialViewController()
         window?.rootViewController = SplashViewController()
         window?.makeKeyAndVisible()
     }
