+++ ImageFeed/Service/ProfileImageService.swift
@@ -9,20 +9,24 @@ import Foundation
 
 final class ProfileImageService {
     
+    //MARK: - Private Properties
+    
     static let didChangeNotification = Notification.Name(rawValue: "ProfileImageProviderDidChange")
     
     static let shared = ProfileImageService()
     private init() {}
     
+    private let tokenStoreg = OAuth2TokenStorage.shared
+    
     private (set) var avatarURL: String?
     private var task: URLSessionTask?
-    private let tokenStoreg = OAuth2TokenStorage.shared
+    
+    //MARK: - Lifecycle
     
     func fetchProfileImageURL(username: String, _ completion: @escaping (Result<String, Error>) -> Void) {
         guard let request = makeFetchProfileImageURLRequest(username: username) else {
             assertionFailure("Invalid request")
             completion(.failure(NetworkError.invalidRequest))
-            
             return
         }
         
@@ -34,7 +38,6 @@ final class ProfileImageService {
                 switch response {
                 case .success(let userResult):
                     self?.handleSuccess(userResult: userResult, completion: completion)
-
                 case .failure(let error):
                     print("Ошибка сетевого запроса в функции \(#function): \(error.localizedDescription)")
                     completion(.failure(error))
@@ -44,19 +47,21 @@ final class ProfileImageService {
         task.resume()
     }
     
+    //MARK: - Private Lifecycle
+    
     private func handleSuccess(userResult: UserResult, completion: (Result<String, Error>) -> Void) {
-            avatarURL = userResult.profileImage.large
-            NotificationCenter.default.post(
-                name: Self.didChangeNotification,
-                object: self,
-                userInfo: ["URL": avatarURL as Any])
-            completion(.success(userResult.profileImage.large))
-        }
+        avatarURL = userResult.profileImage.large
+        NotificationCenter.default.post(
+            name: Self.didChangeNotification,
+            object: self,
+            userInfo: ["URL": avatarURL as Any])
+        completion(.success(userResult.profileImage.large))
+    }
     
     private func makeFetchProfileImageURLRequest(username: String) -> URLRequest? {
         guard let url = URL(string: Constants.profilePublicURLString + username) else { return nil }
         guard let token = tokenStoreg.token else { return nil }
-
+        
         var request = URLRequest(url: url)
         request.httpMethod = "GET"
         request.addValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
